+++ test_flask_ldapconn.py
@@ -9,7 +9,9 @@ import string
 import unittest
 import flask
 
-from ldap3 import SUBTREE, LDAPEntryError
+from ldap3 import SUBTREE, STRING_TYPES
+from ldap3.core.exceptions import LDAPAttributeError
+
 from flask_ldapconn import LDAPConn
 
 from flask_ldapconn.entry import LDAPEntry
@@ -65,6 +67,18 @@ class Account(User):
     home = LDAPAttribute('homeDirectory')
     password = LDAPAttribute('userPassword')
 
+class Account(User):
+    # LDAP meta-data
+    object_classes = ['posixAccount']
+
+    # posixAccount
+    uidnumber = LDAPAttribute('uidNumber')
+    gidnumber = LDAPAttribute('gidNumber')
+    shell = LDAPAttribute('loginShell')
+    home = LDAPAttribute('homeDirectory')
+    password = LDAPAttribute('userPassword')
+
+
 class LDAPConnTestCase(unittest.TestCase):
 
     def setUp(self):
@@ -117,7 +131,7 @@ class LDAPConnModelTestCase(unittest.TestCase):
             entry = self.user.query.filter(
                 'email: %s' % self.app.config['USER_EMAIL']
             ).first()
-            self.assertEqual(entry.email.value,
+            self.assertEqual(entry.email,
                              self.app.config['USER_EMAIL'])
 
     def test_model_search_set_attribute(self):
@@ -127,7 +141,7 @@ class LDAPConnModelTestCase(unittest.TestCase):
                 'email: %s' % self.app.config['USER_EMAIL']
             ).first()
             entry.email = new_email
-            self.assertEqual(entry.email.value, new_email)
+            self.assertEqual(entry.email, new_email)
 
     def test_model_search_set_attribute_list(self):
         new_email_list = ['philip@planetexpress.com',
@@ -137,25 +151,25 @@ class LDAPConnModelTestCase(unittest.TestCase):
                 'email: %s' % self.app.config['USER_EMAIL']
             ).first()
             entry.email = new_email_list
-            self.assertEqual(entry.email.value, new_email_list)
+            self.assertEqual(entry.email, new_email_list)
 
     def test_model_search_set_undefined_attr(self):
         def new_model():
             user = self.user(active='1')
         with self.app.test_request_context():
-            self.assertRaises(LDAPEntryError, new_model)
+            self.assertRaises(LDAPAttributeError, new_model)
 
     def test_model_new(self):
         with self.app.test_request_context():
             user = self.user(name='Rafael Römhild',
                              email='rafael@planetexpress.com')
-            self.assertEqual(user.email.value, 'rafael@planetexpress.com')
+            self.assertEqual(user.email, 'rafael@planetexpress.com')
 
     def test_model_fetch_entry(self):
         uid = 'bender'
         with self.app.test_request_context():
             user = self.user.query.filter('userid: {}'.format(uid)).first()
-            self.assertEqual(user.userid.value, uid)
+            self.assertEqual(user.userid, uid)
 
     def test_model_fetch_entry_with_components_in_and_false(self):
             uid = 'bender'
@@ -163,7 +177,7 @@ class LDAPConnModelTestCase(unittest.TestCase):
                 user = self.user.query.filter(
                     'email: {0}, userid: {0}'.format(uid)
                 ).all(components_in_and=False)
-                self.assertEqual(user[0].userid.value, uid)
+                self.assertEqual(user[0].userid, uid)
 
     def test_model_fetch_entry_authenticate(self):
         uid = 'fry'
@@ -186,7 +200,7 @@ class LDAPConnModelTestCase(unittest.TestCase):
         with self.app.test_request_context():
             entries = self.user.query.filter(query_filter).all()
             for entry in entries:
-                response_uids.append(entry.userid.value)
+                response_uids.append(entry.userid)
         matched_uids = set(expected_uids).intersection(response_uids)
         self.assertEqual(len(expected_uids), len(matched_uids))
 
@@ -200,7 +214,7 @@ class LDAPConnModelTestCase(unittest.TestCase):
         dn = 'cn=Amy Wong+sn=Kroker,ou=people,dc=planetexpress,dc=com'
         with self.app.test_request_context():
             user = self.user.query.get(dn)
-            self.assertEqual('Kroker', user.surname.value)
+            self.assertEqual('Kroker', user.surname)
 
     def test_model_get_attributes_dict(self):
         with self.app.test_request_context():
@@ -245,17 +259,17 @@ class LDAPConnModelTestCase(unittest.TestCase):
         with self.app.test_request_context():
             user = self.user.query.filter('userid: fry').first()
             user['userid'] = 'xyz'
-            self.assertEqual(user.userid.value, 'xyz')
+            self.assertEqual(user.userid, 'xyz')
 
     def test_model_attribute_str(self):
         with self.app.test_request_context():
             user = self.user.query.filter('userid: fry').first()
-            self.assertTrue(isinstance(user.userid, self.ldap.Attribute))
+            self.assertTrue(isinstance(user.userid, STRING_TYPES))
 
     def test_model_attribute_iter(self):
         with self.app.test_request_context():
             user = self.user.query.filter('userid: professor').first()
-            self.assertTrue(isinstance(user.email.value, list))
+            self.assertTrue(isinstance(user.email, list))
             for mail in user.email:
                 pass
 
@@ -270,7 +284,7 @@ class LDAPConnModelTestCase(unittest.TestCase):
                                  givenname='Raphael')
             self.assertTrue(new_user.save())
             user = self.user.query.filter(query_filter).first()
-            self.assertEqual(new_user.userid.value, user.userid.value)
+            self.assertEqual(new_user.userid, user.userid)
 
     def test_model_operation_modify(self):
         uid = 'rafael-{}'.format(UID_SUFFIX)
@@ -279,12 +293,67 @@ class LDAPConnModelTestCase(unittest.TestCase):
             mod_user = self.user.query.filter(query_filter).first()
             mod_user.givenname = 'Rafael'
             mod_user.title = 'SysAdmin'
-            mod_user.email.append('it@planetexpress.co')
+            mod_user.email = [mod_user.email, 'it@planetexpress.co']
             self.assertTrue(mod_user.save())
             user = self.user.query.filter(query_filter).first()
-            self.assertEqual(user.givenname.value, 'Rafael')
-            self.assertEqual(user.surname.value, u'Römhild')
-            self.assertEqual(user.title.value, 'SysAdmin')
+            self.assertEqual(user.givenname, 'Rafael')
+            self.assertEqual(user.surname, u'Römhild')
+            self.assertEqual(user.title, 'SysAdmin')
+            self.assertTrue('it@planetexpress.co' in user.email)
+
+    def test_model_operation_remove(self):
+        uid = 'rafael-{}'.format(UID_SUFFIX)
+        query_filter = 'userid: {}'.format(uid)
+        with self.app.test_request_context():
+            user = self.user.query.filter(query_filter).first()
+            user.delete()
+            user = self.user.query.filter(query_filter).first()
+            self.assertEqual(user, None)
+
+
+class LDAPConnModelInheritanceTestCase(unittest.TestCase):
+
+    def setUp(self):
+        app = flask.Flask(__name__)
+        app.config.from_object(__name__)
+        app.config.from_envvar('LDAP_SETTINGS', silent=True)
+        ldap = LDAPConn(app)
+
+        self.app = app
+        self.ldap = ldap
+        self.user = Account
+
+    def test_model_operation_add(self):
+        uid = 'rafael-{}'.format(UID_SUFFIX)
+        query_filter = 'userid: {}'.format(uid)
+        with self.app.test_request_context():
+            new_user = self.user(name='Rafael Römhild',
+                                 userid=uid,
+                                 email='rafael@planetexpress.com',
+                                 surname='Römhild',
+                                 givenname='Raphael',
+                                 uidnumber=1000,
+                                 gidnumber=1000,
+                                 shell='/bin/false',
+                                 home='/home/' + uid)
+            self.assertTrue(new_user.save())
+            user = self.user.query.filter(query_filter).first()
+            self.assertEqual(new_user.userid, user.userid)
+
+    def test_model_operation_modify(self):
+        uid = 'rafael-{}'.format(UID_SUFFIX)
+        query_filter = 'userid: {}'.format(uid)
+        with self.app.test_request_context():
+            mod_user = self.user.query.filter(query_filter).first()
+            mod_user.givenname = 'Rafael'
+            mod_user.title = 'SysAdmin'
+            mod_user.shell = '/bin/bash'
+            mod_user.email = [mod_user.email, 'it@planetexpress.co']
+            self.assertTrue(mod_user.save())
+            user = self.user.query.filter(query_filter).first()
+            self.assertEqual(user.givenname, 'Rafael')
+            self.assertEqual(user.surname, u'Römhild')
+            self.assertEqual(user.title, 'SysAdmin')
             self.assertTrue('it@planetexpress.co' in user.email)
 
     def test_model_operation_remove(self):